refactor: move isBrowser utility to parent directory
