+++ test/playground/pages/isBrowser.ts
